Revert "Consider range tombstone in compaction output file cutting (#10802)"
This reverts commit f02c708aa32829bbbd70aa3493af8444e76e4350.
